Wrap .org XML RPC error `enum` in an `Error`-conforming `struct`
Jumping through this hoop was necessary to avoid the Swift compiler
automatically generating an error domain for the `Error` and making it
impossible to successfully redefine the domain at the `CustomNSError`
conformance site.
